starting to implement scene switching; implementing exp, level up and move learning; finalising basic battle system

+++ PokemonRevolution/Assets/Scripts/Battle/AttackInfo.cs
@@ -1,11 +1,11 @@
 public struct AttackInfo
 {
     public bool criticalHit;
-    public int damage;
+    public float damage;
     public float typeEffectiveness;
     public bool moveHits;
 
-    public AttackInfo(bool criticalHit, int damage, float typeEffectiveness, bool moveHits)
+    public AttackInfo(bool criticalHit, float damage, float typeEffectiveness, bool moveHits)
     {
         this.criticalHit = criticalHit;
         this.damage = damage;
